backend/app/requirements.txt
fastapi
uvicorn
opencv-python
python-multipart
numpy
openai
sentence-transformers
faiss-cpu
pydantic
python-dotenv
